9/16 lessons from nextjs.org
